refactor(plugin): extract compute_kernel.rs (#1079)
Moves the VulkanComputeKernelMethodsVTable callbacks +
HOST_VULKAN_COMPUTE_KERNEL_METHODS_VTABLE + accessor +
compute_kernel_methods_vtable_null_tests +
kernel_bindings_vtable_tier1_wire_format_tests (the cross-kernel
bindings tier-1 module lands here since compute is the canonical
kernel; the graphics/ray-tracing branches reach the other vtables
via super::super::HOST_VULKAN_GRAPHICS_KERNEL_METHODS_VTABLE /
super::super::HOST_VULKAN_RAY_TRACING_KERNEL_METHODS_VTABLE for now).

mod.rs re-exports HOST_VULKAN_COMPUTE_KERNEL_METHODS_VTABLE and
host_vulkan_compute_kernel_methods_vtable.

cargo test -p streamlib-engine --lib unchanged at 1407 passed / 2
pre-existing Deno failures / 126 ignored.
